formulas update
