+++ app/Http/Traits/AnalyticsTrait.php
@@ -11,11 +11,10 @@ trait AnalyticsTrait
      * @param float|int $marketing_cost marketing_costs за месяц
      * @return float|int revenue needed
      */
-    private function getRevenueNeeded($marketing_cost)
+    public function getRevenueNeeded($marketing_cost)
     {
-//        dd($this->globalcogs);
         if ($this->duration > 1) {
-            return $this->fixed_costs / ($this->globalcogs - $marketing_cost);
+            return $this->fixed_costs / (1 - $this->globalcogs - $marketing_cost);
         }
 
         return $this->fixed_costs / (1 - $this->globalcogs - $marketing_cost);
